fix(telemetry): full-precision heading/yaw (f64) + north-spin + frozen FPM
- Store/carry yaw (FC fused heading) and heading (course over ground) as f64 with
  full precision instead of rounded i16, across the whole pipeline: TelemetryRecord +
  AttitudeData, MSP/MAVLink/FrSky live, Blackbox/ArduPilot/raw import, recorder, DB
  (read via inference, write via ToSql; SQLite stores REAL via affinity — no schema
  migration), replay and CSV export.
- Blackbox: yaw (attitude[2], decidegrees) now ALWAYS ÷10 — the old per-value
  ">360 → ÷10" heuristic left the 0–36° band undivided, causing a heading spin on
  every north crossing. COG (gps_ground_course, degrees) kept as-is (>360 fallback).
- Flight-path marker (AHI + FPV HUD): ease in a reactive $effect (untracked self-read)
  instead of a requestAnimationFrame loop — the rAF read a reactive value that went
  stale and pinned the marker at the dial edge. Drop the now-unused easeFactor helper.

+++ src-tauri/src/flightlog/recorder.rs
@@ -277,13 +277,13 @@ struct TelemetrySnapshot {
     // Attitude
     roll: Option<f64>,
     pitch: Option<f64>,
-    yaw: Option<i16>,
+    yaw: Option<f64>,
     // GPS
     lat: Option<f64>,
     lon: Option<f64>,
     alt_gps: Option<f64>,
     speed: Option<f64>,
-    heading: Option<i16>,
+    heading: Option<f64>,
     fix_type: Option<u8>,
     num_sat: Option<u8>,
     // Altitude (baro)
@@ -493,7 +493,7 @@ impl FlightRecorder {
         self.snapshot.speed = Some(data.ground_speed);
         // DB column `heading` = course over ground (GpsData.course); the FC fused heading is stored
         // separately in `yaw` (from on_attitude). Kept distinct for the wind/crab analysis.
-        self.snapshot.heading = Some(data.course as i16);
+        self.snapshot.heading = Some(data.course);
         self.snapshot.fix_type = Some(data.fix_type);
         self.snapshot.num_sat = Some(data.num_sat);
         self.maybe_record_sample();
